pyproject.toml
[project]
name = "orthoai-agent"
version = "0.1.0"
description = "Medical context agents for AgentBeats evaluation platform"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "a2a-sdk[http-server]>=0.3.8",
    "dotenv>=0.9.9",
    "typer>=0.19.2",
    "uvicorn>=0.37.0",
    "httpx>=0.27.0",
    "litellm>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

